change doctor reg
change doctor reg
